Yoink pg_stat_wal collector (#10)
A `postgres_exporter` contributor created a pull request to add a
collector for `pg_stat_wal` but this was never merged by the project:
https://github.com/prometheus-community/postgres_exporter/pull/1082

I want this for some dashboard work, so I yoinked it, fixed up a handful
of linter errors and pushed it.

I've tested the changes locally in docker compose and they work.

---------

Signed-off-by: Harry Cho <pandar00@hotmail.com>
Signed-off-by: Harry Cho <pandar00@Hotmail.com>
Co-authored-by: Harry Cho <pandar00@hotmail.com>
Co-authored-by: Ben Kochie <superq@gmail.com>